Make Janet version a build time are
Janet versions are released quite often. This changes helps building
new images with an updated version of janet without modifying the
Dockerfile.
This can be done with:

```
docker build --build-arg JANET=1.12.1 -f docker/Dockerfile .
```

If JANET isn't passed as a build-arg the default value is used. This
is currently set to 1.12.2. Which is the the latest janet version.